requirements.txt
fastapi==0.115.*
uvicorn[standard]==0.30.*
jinja2==3.*
pydantic==2.*
google-api-python-client==2.*
google-auth==2.*
google-auth-httplib2==0.2.*
google-cloud-secret-manager==2.*
cachetools==5.*
pyjwt==2.*
google-cloud-storage==2.*
google-cloud-bigquery==3.*
